irsummary_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
camelot-py[cv]==0.10.1
pdfplumber==0.9.0
pandas==2.1.3
psycopg2-binary==2.9.9
PyMuPDF==1.23.7
httpx==0.26.0
tqdm 
python-multipart==0.0.6
